Atualização e ajuste de campo da tabela

+++ queries.sql
@@ -21,7 +21,7 @@ CREATE TABLE IF NOT EXISTS produtos(
    nome_produto VARCHAR (160),
    modelo VARCHAR (80),
    lote INT (20),
-   preco FLOAT (5, 2),
+   preco DECIMAL (5, 2),
    quantidade INT (20),
    constraint produto_pk PRIMARY KEY AUTO_INCREMENT (id_produto),
    constraint fornecedores_produtos_fk FOREIGN KEY (id_forn)
